Rename existing local date utils file
