+++ apps/website/src/server/trpc/router/admin/notifications.ts
@@ -27,8 +27,8 @@ import {
 
 import { permissions } from "@/data/permissions";
 
+import { inputValueDatetimeLocalToUtc } from "@/utils/datetime-local";
 import { imageMimeTypes } from "@/utils/files";
-import { inputValueDatetimeLocalToUtc } from "@/utils/local-datetime";
 
 const uploadPrefix = "notifications/";
